refactor: 배럴패턴 및 절대경로 수정(#38)
- 컨벤션에 따른 배럴패턴 및 절대경로 수정
