--- src/components/common/dropdown/index.ts
@@ -1,5 +1,0 @@
-export { default as Dropdown } from './Dropdown';
-export { default as DropdownContent } from './DropdownContent';
-export { default as DropdownItem } from './DropdownItem';
-export { default as DropdownSeparator } from './DropdownSeparator';
-export { default as DropdownTrigger } from './DropdownTrigger';
