+++ src/components/common/modal/ModalTrigger.tsx
@@ -1,4 +1,4 @@
-import { DialogTrigger } from '@/components/ui/dialog';
+import { DialogTrigger } from '@/components';
 
 type ModalTriggerProps = {
   children: React.ReactNode;
