--- src/components/common/modal/index.ts
@@ -1,7 +1,0 @@
-export { default as Modal } from './Modal';
-export { default as ModalContent } from './ModalContent';
-export { default as ModalDescription } from './ModalDescription';
-export { default as ModalFooter } from './ModalFooter';
-export { default as ModalHeader } from './ModalHeader';
-export { default as ModalTitle } from './ModalTitle';
-export { default as ModalTrigger } from './ModalTrigger';
