+++ src/components/index.ts
@@ -1,2 +1,28 @@
-export { default as MainLayout } from '@/components/layout/MainLayout';
-export { Button, buttonVariants } from '@/components/ui/button';
+//dropdown
+export { default as Dropdown } from './common/dropdown/Dropdown';
+export { default as DropdownContent } from './common/dropdown/DropdownContent';
+export { default as DropdownItem } from './common/dropdown/DropdownItem';
+export { default as DropdownSeparator } from './common/dropdown/DropdownSeparator';
+export { default as DropdownTrigger } from './common/dropdown/DropdownTrigger';
+//modal
+export { default as Modal } from './common/modal/Modal';
+export { default as ModalContent } from './common/modal/ModalContent';
+
+export { default as ModalDescription } from './common/modal/ModalDescription';
+export { default as ModalFooter } from './common/modal/ModalFooter';
+export { default as ModalHeader } from './common/modal/ModalHeader';
+export { default as ModalTitle } from './common/modal/ModalTitle';
+export { default as ModalTrigger } from './common/modal/ModalTrigger';
+
+//layout
+export { default as Header } from './layout/Header';
+export { default as MainLayout } from './layout/MainLayout';
+export { default as Sidebar } from './layout/Sidebar';
+
+export { Button, buttonVariants } from './ui/Button';
+//ui
+export * from './ui/Dialog';
+export * from './ui/DropdownMenu';
+export { Input } from './ui/Input';
+export { Loading } from './ui/Loading';
+export { Toaster } from './ui/Sonner';
