+++ src/components/layout/Header.tsx
@@ -1,15 +1,14 @@
 import { MenuIcon } from 'lucide-react';
 import { Link } from 'react-router';
 
+import { Button } from '@/components';
 import { ROUTES_PATHS } from '@/constants';
-
-import { Button } from '../ui/button';
 
 type HeaderProps = {
   onMenuClick: () => void;
 };
 
-export function Header({ onMenuClick }: HeaderProps) {
+export default function Header({ onMenuClick }: HeaderProps) {
   return (
     <header className="sticky top-0 z-60 flex items-center justify-between border-b border-[#E5E7EB] bg-[#070913] px-6 py-3">
       <div className="flex items-center justify-start gap-3">
@@ -34,5 +33,3 @@ export function Header({ onMenuClick }: HeaderProps) {
     </header>
   );
 }
-
-export default Header;
