+++ src/components/layout/MainLayout.tsx
@@ -1,10 +1,7 @@
 import { useState } from 'react';
 import { Outlet } from 'react-router';
 
-import { Toaster } from '@/components/ui/sonner';
-
-import { Header } from '../layout/Header';
-import { Sidebar } from '../layout/Sidebar';
+import { Header, Sidebar, Toaster } from '@/components';
 
 export default function MainLayout() {
   const [isOpen, setIsOpen] = useState(false);
