+++ src/components/ui/Dialog.tsx
@@ -2,7 +2,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { XIcon } from 'lucide-react';
 import * as React from 'react';
 
-import { cn } from '@/utils/index';
+import { cn } from '@/utils';
 
 function Dialog({
   ...props
