+++ src/constants/index.ts
@@ -1,2 +1,2 @@
-export * from '@/constants/colors';
-export * from '@/constants/urls';
+export * from './colors';
+export * from './urls';
