+++ src/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
 
-import { handlers } from '@/mocks/handlers';
+import { handlers } from './handlers';
 
 export const worker = setupWorker(...handlers);
